Added Deferred events for animation race condition
Tweens are great and all, however since the tween expect run
inbetween certain event updates, there are chances of operations
on reassisnged or null objects when the tween reaches a certain state

Most notable when trying to get dead tiles when adding new ones back
to the board. A deferred event will check the IDLE state of the
display board for active tweens and add itself to an event list.
This gets checked after tweens complete to pump through waiting
sets of tweens.

Another solution may be to only deactivate objects once their tweens
are complete